feat: change the API of `findDevice` to return null if no device is found, rather than throwing an error
BREAKING CHANGE: This refactors the API of the `findDevice`
function to return `null` rather than throwing an error if a
Logitech Litra Glow device cannot be found attached to the
computer. If you're using this function, you should update any
call sites so they don't assume that a device has been returned -
for example with an `if (device)` `null` check.