feat: update command line utilities to enable targeting a specific device with `--serial-number`/`-s`
This updates our command line utilities - `litra-on`, `litra-off`,
`litra-brightness`, `litra-brightness-lm`, `litra-temperature` and
`litra-temperature-k` - to support targeting a specific device
with `--serial-number`/`-s`.

This is useful if you have multiple Litra devices. If you have
multiple devices and don't specify a serial number, as before,
the code will pick whatever happens to be the first device, and
it will print a warning suggesting that you use the option.

As part of this change, we also switch to using Commander for
command line argument parsing, which gives us a `--help` command
for free.