feat: add new `litra-identify` command for identifying the serial number of Litra devices
In 323978c, we add support for pointing to a specific Litra device
when running our command line tools, which is useful if you have
multiple devices. The tricky bit is knowing the serial number of
your devices!

This adds a new `litra-identify` command - which includes a `--help`
command thanks to Commander - which runs through each of your
Litra devices, flashing each one on and off and printing the serial
number so you can know which device is which!