fix: remove the `litra-temperature` CLI, because you can't actually set a Litra's temperature to an arbitrary percentage
BREAKING CHANGE: The `litra-temperature` CLI has been removed
because it claims to allow users to set their Litra device's
temperature toan arbitrary percentage value, but the device only
allows a limited set of temperatures (multiples of 100 between the
minimum and maximum) - see #92. The `litra-temperature-k` CLI for
setting the temperature to a value in Kelvin is still available.