fix: remove the `setTemperaturePercentage` function, because you can't actually set a Litra's temperature to an arbitrary percentage
BREAKING CHANGE: The `setTemperaturePercentage` function has been
removed because it claims to allow users to set their Litra device's
temperature toan arbitrary percentage value, but the device only
allows a limited set of temperatures (multiples of 100 between the
minimum and maximum) - see #92. The `setTemperatureInKelvin`
function for setting the temperature to a value in Kelvin is still
available.