fix: throw an error if `setTemperatureInKelvin` is called with a value that isn't a multiple of 100
BREAKING CHANGE: `setTemperatureInKelvin` will now throw an
error if called with a value that is within the device's supported
temperature range, but which is not a multiple of 100. Currently,
`setTemperatureInKelvin` can be called with any integer
value within the device's allowed temperature range, but in
practice, this does not work. If you try to use a value that isn't
a multiple of 100, the function does not throw an error, but your
device's temperature does not change.

Fixes #92.