Fix `litra-oversight`'s compatibility with Oversight 2.1.8+ by accepting `-activeCount` argument
`litra-oversight` allows you to automatically turn your Litra
on/off when your webcam turns on and off in connection with
[OverSight][1].

OverSight can be pointed at the `litra-oversight` command to
enable this kind of automation.

OverSight calls the specified CLI with a number of arguments.
v2.1.8 [added][2] a new `-activeCount` argument that is
passed. We don't actually need this or care about it, but
since it isn't an allowlisted argument, we error and the
integration breaks down.

This allows `litra-oversight` to accept but ignore that
argument, getting thingss working again.

Fixes #259.

[1]: https://github.com/objective-see/OverSight
[2]: https://github.com/objective-see/OverSight/pull/48